Refactor frontend deployment preset and introduce buildman
